Move container validation to DxilContainerValidation.cpp.
