+++ lib/DxilValidation/CMakeLists.txt
@@ -4,7 +4,9 @@ add_hlsl_hctgen(DxilValidationInc OUTPUT DxilValidation.inc BUILD_DIR)
 add_hlsl_hctgen(DxilValidation OUTPUT DxilValidationImpl.inc BUILD_DIR)
 
 add_llvm_library(LLVMDxilValidation
+  DxilContainerValidation.cpp
   DxilValidation.cpp
+  DxilValidationUtils.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${LLVM_MAIN_INCLUDE_DIR}/llvm/IR
